Extracted getJwtToken method into Utilities class
